chore: move Write components to write folder
