+++ src/components/write/category-button.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import Button from './ui/button';
+import Button from '../button';
 import styled from 'styled-components';
 import { useState } from 'react';
