+++ src/pages/write-post.jsx
@@ -1,7 +1,7 @@
 import styled from "styled-components";
 
-import WriteForm from "../components/write-form";
-import CategoryBar from "../components/category-bar";
+import WriteForm from "../components/write/write-form";
+import CategoryBar from "../components/ui/write/category-bar";
 
 const Wrapper = styled.div`
     margin: 50px 150px;
